refactor(cli): split session command responsibilities
