+++ apps/cli/src/commands/session/summary.ts
