+++ apps/cli/test/integration/session-command.test.ts
@@ -10,8 +10,11 @@ import { StationSnapshotSchema } from "@station/contracts";
 import type { TerminalCommandRecord } from "@station/protocol";
 import { describe, expect, it, vi } from "vitest";
 import { createTempState, writeConfigToml } from "../../../../tests/support/temp-projects";
-import { runSessionCommand, sessionCommandExitCode } from "../../src/commands/session.js";
-import { renderSessionCommandText } from "../../src/commands/sessionOutput.js";
+import {
+  renderSessionCommandText,
+  runSessionCommand,
+  sessionCommandExitCode,
+} from "../../src/commands/session/index.js";
 
 const now = "2026-08-21T12:00:00.000Z";
 const observerBuildVersion = `0.0.0-local+station.${"a".repeat(64)}`;
